# server/src/main/java/kanban/board/server/utils/ApplicationResponse.java
package kanban.board.server.utils;

import lombok.AllArgsConstructor;
import lombok.Getter;

@AllArgsConstructor
@Getter
public class ApplicationResponse {

    private String response;

    public ApplicationResponse(ErrorCodes errorCodes){
        this.response = errorCodes.getResponse();
    }
}